feat: Implement address service with address resolution and coordinate retrieval
feat: Add inventory service for inventory validation, reservation, and restoration

feat: Create location service to find nearest store based on coordinates

feat: Develop order creation service with inventory validation and location resolution

feat: Implement order query service for listing and retrieving orders

feat: Add order status service for managing order shipping, confirmation, and cancellation
