+++ apps/api/src/controllers/order.controller.ts
@@ -1,15 +1,16 @@
 import { Request, Response } from "express";
 import type { Request as ExpressRequest } from "express";
-// Reuse the AuthRequest shape used in auth middleware
-type AuthRequest = ExpressRequest & { user?: { id: number } };
 import { OrderService } from "../services/order.service.js";
-// cleaned unused imports
+import { orderQueryService } from "../services/order-query.service.js";
 import { prisma } from "@repo/database";
 import {
   successResponse,
   errorResponse,
   ERROR_MESSAGES,
 } from "../utils/helpers.js";
+
+// Reuse the AuthRequest shape used in auth middleware
+type AuthRequest = ExpressRequest & { user?: { id: number } };
 
 function toNumber(value: unknown, fallback?: number) {
   if (value === undefined || value === null) return fallback;
@@ -39,10 +40,16 @@ function pickUserId(req: Request): number | undefined {
   return undefined;
 }
 
+/**
+ * Order Controller - Handles HTTP requests for order operations
+ * Uses the refactored OrderService which delegates to specialized services
+ */
 export class OrderController {
   private service = new OrderService();
 
-  // GET /orders - list with filters
+  /**
+   * GET /orders - List orders with filters and pagination
+   */
   listOrders = async (req: Request, res: Response) => {
     try {
       const userId = pickUserId(req);
@@ -82,29 +89,42 @@ export class OrderController {
     }
   };
 
+  /**
+   * POST /orders - Create a new order
+   */
   createOrder = async (req: Request, res: Response) => {
     try {
       const userId = pickUserId(req);
-      // allow storeId to be omitted so backend can compute nearest store
+      if (!userId) {
+        return res.status(400).json(errorResponse("Missing userId in request"));
+      }
+
+      // Extract request parameters
       const storeId = toNumber(req.body?.storeId, undefined);
       const items = req.body?.items;
       const userLat = req.body?.userLat ? Number(req.body.userLat) : undefined;
       const userLon = req.body?.userLon ? Number(req.body.userLon) : undefined;
-      const addressId = req.body?.addressId
-        ? Number(req.body.addressId)
-        : undefined;
-      // TODO: Accept and forward an idempotency key so retries from the client
-      // (or network retries) don't create duplicate orders. Replace in-memory
-      // strategy with DB-backed idempotency table in production.
+      const addressId = req.body?.addressId ? Number(req.body.addressId) : undefined;
+      
+      // Idempotency key handling
       const idempotencyKey =
         (req.headers["idempotency-key"] as string) ||
         req.body?.idempotencyKey ||
         req.query?.idempotencyKey;
 
-      if (!userId)
-        return res.status(400).json(errorResponse("Missing userId in request"));
-      if (!Array.isArray(items) || items.length === 0)
+      // Validate items
+      if (!Array.isArray(items) || items.length === 0) {
         return res.status(400).json(errorResponse("Missing items in order"));
+      }
+
+      // Validate item structure
+      for (const item of items) {
+        if (!item.productId || !item.qty || item.qty <= 0) {
+          return res.status(400).json(
+            errorResponse("Invalid item structure", "Each item must have productId and positive qty")
+          );
+        }
+      }
 
       const result = await this.service.createOrder(
         userId,
@@ -115,56 +135,27 @@ export class OrderController {
         userLon,
         addressId
       );
+
       return res.status(201).json(successResponse(result, "Order created"));
     } catch (e) {
-      const msg = e instanceof Error ? e.message : String(e);
-
-      // Map known inventory errors to 400 so frontend can surface actionable UX
-      if (
-        msg.includes(ERROR_MESSAGES.INVENTORY.INSUFFICIENT_STOCK) ||
-        msg.includes(ERROR_MESSAGES.INVENTORY.NO_INVENTORY) ||
-        msg.includes(ERROR_MESSAGES.STORE.NO_NEARBY) ||
-        msg.toLowerCase().includes("not found")
-      ) {
-        // Provide a concise client-friendly message and include the raw error
-        // in `error` for debugging and an `errors` entry for field-level hints.
-        return res
-          .status(400)
-          .json(
-            errorResponse("Order cannot be created", msg, [
-              { field: "items", message: msg },
-            ])
-          );
-      }
-
-      return res.status(500).json(errorResponse("Failed to create order", msg));
+      return this._handleOrderError(e, res, "Failed to create order");
     }
   };
 
+  /**
+   * GET /orders/:id - Get order by ID
+   */
   getOrderById = async (req: Request, res: Response) => {
     try {
       const id = Number(req.params.id);
-      if (!id) return res.status(400).json(errorResponse("Invalid order id"));
+      if (!id) {
+        return res.status(400).json(errorResponse("Invalid order id"));
+      }
 
-      // Use prisma directly for a simple read operation including relations.
-      // Include product.name for each order item so frontend can render the
-      // human-friendly product name instead of a fallback like "Product #3".
-      const order = await prisma.order.findUnique({
-        where: { id },
-        include: {
-          items: {
-            include: {
-              product: {
-                select: { id: true, name: true, price: true },
-              },
-            },
-          },
-          payment: true,
-          shipment: true,
-        },
-      });
-
-      if (!order) return res.status(404).json(errorResponse("Order not found"));
+      const order = await orderQueryService.getOrderById(id);
+      if (!order) {
+        return res.status(404).json(errorResponse("Order not found"));
+      }
 
       // Ownership check: allow request if same user or in dev fallback
       const userId = pickUserId(req);
@@ -183,31 +174,39 @@ export class OrderController {
     }
   };
 
+  /**
+   * POST /orders/:id/payment-proof - Upload payment proof
+   */
   uploadPaymentProof = async (req: Request, res: Response) => {
     try {
       const id = Number(req.params.id);
-      if (!id) return res.status(400).json(errorResponse("Invalid order id"));
-      // Expect multipart/form-data with a single file field named 'proof'
+      if (!id) {
+        return res.status(400).json(errorResponse("Invalid order id"));
+      }
+
+      // Validate file upload
       const f = (req as any).file as
         | { buffer?: Buffer; mimetype?: string }
         | undefined;
-      if (!f || !f.buffer)
+      if (!f || !f.buffer) {
         return res
           .status(400)
           .json(errorResponse("Missing file 'proof' in request"));
+      }
 
+      // Validate file type
       const allowed = ["image/png", "image/jpeg", "image/jpg"];
       const mime = f.mimetype ?? "application/octet-stream";
-      if (!allowed.includes(mime))
+      if (!allowed.includes(mime)) {
         return res.status(400).json(errorResponse("Invalid file type"));
+      }
 
-      const MAX_BYTES = 1 * 1024 * 1024;
-      if (f.buffer.length > MAX_BYTES)
+      // Validate file size
+      const MAX_BYTES = 1 * 1024 * 1024; // 1MB
+      if (f.buffer.length > MAX_BYTES) {
         return res.status(400).json(errorResponse("File too large"));
+      }
 
-      // Hand off buffer and mime to service which will upload to Cloudinary and return
-      // proof URL plus the updated payment and order status so the frontend
-      // can update UI without an immediate refetch.
       const uploadResult = await this.service.uploadPaymentProof(
         id,
         f.buffer,
@@ -227,82 +226,80 @@ export class OrderController {
     }
   };
 
+  /**
+   * PATCH /orders/:id/cancel - Cancel order
+   */
   cancelOrder = async (req: Request, res: Response) => {
     try {
       const id = Number(req.params.id);
-      if (!id) return res.status(400).json(errorResponse("Invalid order id"));
+      if (!id) {
+        return res.status(400).json(errorResponse("Invalid order id"));
+      }
 
       const userId = pickUserId(req);
-      if (!userId)
+      if (!userId) {
         return res.status(400).json(errorResponse("Missing userId in request"));
+      }
 
-      // Delegate to service which will perform ownership/status checks
       const result = await this.service.cancelOrder(id, userId);
-
       return res.status(200).json(successResponse(result, "Order cancelled"));
     } catch (e) {
-      const msg = e instanceof Error ? e.message : String(e);
-      if (msg.includes("not found")) {
-        return res.status(404).json(errorResponse("Order not found", msg));
-      }
-      if (msg.includes("Cannot cancel") || msg.includes("already")) {
-        return res.status(409).json(errorResponse("Cannot cancel order", msg));
-      }
-      return res.status(500).json(errorResponse("Failed to cancel order", msg));
+      return this._handleOrderStatusError(e, res, "Failed to cancel order");
     }
   };
 
+  /**
+   * PATCH /orders/:id/confirm - Confirm order receipt
+   */
   confirmOrder = async (req: Request, res: Response) => {
     try {
       const id = Number(req.params.id);
-      if (!id) return res.status(400).json(errorResponse("Invalid order id"));
+      if (!id) {
+        return res.status(400).json(errorResponse("Invalid order id"));
+      }
 
       const userId = pickUserId(req);
-      if (!userId)
+      if (!userId) {
         return res.status(400).json(errorResponse("Missing userId in request"));
+      }
 
       const result = await this.service.confirmOrder(id, userId);
       return res.status(200).json(successResponse(result, "Order confirmed"));
     } catch (e) {
-      const msg = e instanceof Error ? e.message : String(e);
-      if (msg.includes("not found"))
-        return res.status(404).json(errorResponse("Order not found", msg));
-      if (msg.includes("Cannot confirm") || msg.includes("already"))
-        return res.status(409).json(errorResponse("Cannot confirm order", msg));
-      return res.status(500).json(errorResponse("Failed to confirm order", msg));
+      return this._handleOrderStatusError(e, res, "Failed to confirm order");
     }
   };
 
+  /**
+   * PATCH /orders/:id/ship - Mark order as shipped (admin)
+   */
   shipOrder = async (req: Request, res: Response) => {
     try {
       const id = Number(req.params.id);
-      if (!id) return res.status(400).json(errorResponse("Invalid order id"));
+      if (!id) {
+        return res.status(400).json(errorResponse("Invalid order id"));
+      }
 
       const userId = pickUserId(req);
-      // allow worker/admin calls without userId in dev but prefer one when present
       const result = await this.service.shipOrder(id, userId);
       return res.status(200).json(successResponse(result, "Order shipped"));
     } catch (e) {
-      const msg = e instanceof Error ? e.message : String(e);
-      if (msg.includes("not found"))
-        return res.status(404).json(errorResponse("Order not found", msg));
-      if (msg.includes("Cannot ship") || msg.includes("already"))
-        return res.status(409).json(errorResponse("Cannot ship order", msg));
-      return res.status(500).json(errorResponse("Failed to ship order", msg));
+      return this._handleOrderStatusError(e, res, "Failed to ship order");
     }
   };
 
+  /**
+   * GET /orders/counts - Get order counts by status
+   */
   getOrderCounts = async (req: Request, res: Response) => {
     try {
       const userId = pickUserId(req);
-      if (!userId)
+      if (!userId) {
         return res.status(400).json(errorResponse("Missing userId in request"));
+      }
 
       console.log("Getting order counts for userId:", userId);
-
-      // Get counts for each status
       const counts = await this.service.getOrderCountsByStatus(userId);
-
       console.log("Order counts:", counts);
 
       return res
@@ -316,4 +313,46 @@ export class OrderController {
         .json(errorResponse("Failed to get order counts", msg));
     }
   };
+
+  /**
+   * Private helper to handle order creation errors
+   */
+  private _handleOrderError(e: unknown, res: Response, defaultMessage: string) {
+    const msg = e instanceof Error ? e.message : String(e);
+
+    // Map known inventory errors to 400 so frontend can surface actionable UX
+    if (
+      msg.includes(ERROR_MESSAGES.INVENTORY.INSUFFICIENT_STOCK) ||
+      msg.includes(ERROR_MESSAGES.INVENTORY.NO_INVENTORY) ||
+      msg.includes(ERROR_MESSAGES.STORE.NO_NEARBY) ||
+      msg.toLowerCase().includes("not found")
+    ) {
+      return res
+        .status(400)
+        .json(
+          errorResponse("Order cannot be created", msg, [
+            { field: "items", message: msg },
+          ])
+        );
+    }
+
+    return res.status(500).json(errorResponse(defaultMessage, msg));
+  }
+
+  /**
+   * Private helper to handle order status change errors
+   */
+  private _handleOrderStatusError(e: unknown, res: Response, defaultMessage: string) {
+    const msg = e instanceof Error ? e.message : String(e);
+    
+    if (msg.includes("not found")) {
+      return res.status(404).json(errorResponse("Order not found", msg));
+    }
+    
+    if (msg.includes("Cannot") || msg.includes("already")) {
+      return res.status(409).json(errorResponse("Cannot change order status", msg));
+    }
+    
+    return res.status(500).json(errorResponse(defaultMessage, msg));
+  }
 }
