+++ apps/api/src/services/index.ts
@@ -1,0 +1,14 @@
+// Service exports for easy importing
+export { OrderService } from "./order.service.js";
+export { OrderCreationService, orderCreationService } from "./order-creation.service.js";
+export { OrderStatusService, orderStatusService } from "./order-status.service.js";
+export { OrderQueryService, orderQueryService } from "./order-query.service.js";
+export { LocationService, locationService } from "./location.service.js";
+export { InventoryService, inventoryService } from "./inventory.service.js";
+export { AddressService, addressService } from "./address.service.js";
+export { PaymentService, paymentService } from "./payment.service.js";
+export { CartService } from "./cart.service.js";
+export { ProductService } from "./product.service.js";
+
+// Type exports
+export type { OrderListOptions, OrderListResult } from "./order-query.service.js";
